+++ src/auto_drive_sim/src/drive_perception/per_traffic.py
@@ -1,0 +1,54 @@
+#!/usr/bin/env python3
+#-*- coding:utf-8 -*- 
+import sys
+import os
+
+current_dir = os.path.dirname(os.path.abspath(__file__))
+parent_dir = os.path.abspath(os.path.join(current_dir, '..'))  # drive_controller 상위 폴더
+if parent_dir not in sys.path:
+    sys.path.insert(0, parent_dir)
+
+import rospy
+from sensor_msgs.msg import CompressedImage
+from cv_bridge import CvBridge
+import cv2 
+import numpy as np
+from std_msgs.msg import Float64, Bool, String
+from morai_msgs.msg import GetTrafficLightStatus
+import json
+from utils import check_timer
+
+class PerTraffic:
+    def __init__(self):
+        print(f"PerTraffic start")
+        rospy.init_node('per_traffic_node')
+        self.signal = 0
+        self.prev_signal = 0
+        self.traffic_msg = None
+        self.init_pubSub()
+        
+    def init_pubSub(self):
+        rospy.Subscriber("/GetTrafficLightStatus", GetTrafficLightStatus, self.CB_traffic_raw)
+        self.pub_is_go_traffic = rospy.Publisher('/is_to_go_traffic', Bool, queue_size=1)
+
+    def CB_traffic_raw(self, msg):
+        self.traffic_msg = msg
+        if self.traffic_msg.trafficLightIndex == "SN000005":
+            self.signal = self.traffic_msg.trafficLightStatus
+            if self.prev_signal != self.signal:
+                self.prev_signal = self.signal
+                
+    def check_and_publish_signal(self):
+        is_go = self.signal in [16, 33]
+        self.pub_is_go_traffic.publish(Bool(data=is_go))
+
+    def processing(self):
+        rate = rospy.Rate(40)  # 40Hz 루프
+        while not rospy.is_shutdown():
+            # 조건 체크 및 publish
+            self.check_and_publish_signal()
+            rate.sleep()
+            
+if __name__ == '__main__':
+    node = PerTraffic()
+    node.processing()
